Legal Tech/api/pyproject.toml
[tool.poetry]
name = "signature-toolkit-api"
version = "0.1.0"
description = "FastAPI backend for signature extraction toolkit"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
python-multipart = "^0.0.20"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = "^2.11.4"
pydantic-settings = "^2.9.1"
pdfplumber = "^0.10.3"
PyPDF2 = "^3.0.0"
google-generativeai = "^0.3.2"
pytesseract = "^0.3.10"
Pillow = "^10.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"
pytest-asyncio = "^0.26.0"
ruff = "^0.11.8"
black = "^25.1.0"
reportlab = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
select = ["E", "F", "I", "N", "W", "B", "A", "C4", "PT", "SIM", "ARG"]
ignore = ["B008"]
line-length = 100

[tool.black]
line-length = 100
target-version = ["py312"] 